network_daemon/requirements.txt
grpcio
protobuf
util-base
psutil
envoy
lxml
paramiko
